+++ shell/common/exports.sh
@@ -24,4 +24,4 @@ export MANPAGER='less -X';
 
 # Avoid issues with `gpg` as installed via Homebrew.
 # https://stackoverflow.com/a/42265848/96656
-export GPG_TTY=$(tty);
+command -v tty >/dev/null 2>&1 && export GPG_TTY=$(tty);